Add dtype to args and pass args around.

+++ model/eval/evaluator.py
@@ -238,7 +238,7 @@ def evaluate_model_from_name(
     )
 
     model = AutoModelForCausalLM.from_pretrained(
-        model_name, torch_dtype=torch.float16, device_map="auto"
+        model_name, torch_dtype=args.dtype, device_map="auto"
     )
     tokenizer = AutoTokenizer.from_pretrained(model_name, padding_side="left")
